Removed stdint.h dependency and added logging.
Not sure if it's worse or better without stdint.h types, oh well.
Logging is trying to catch a crashing bug that sometimes happens
during manipulator EXTEND with grabbed atom.
Bug has been difficult to reproduce, but my guess is that it has
to do with the the build_compound() process.

+++ atom.h
@@ -1,7 +1,6 @@
 #ifndef _ATOM_H
 #define _ATOM_H
 
-#include <stdint.h>
 #include "space.h"
 #include "list.h"
 
@@ -14,7 +13,7 @@ enum element {
 
 struct atom {
 	struct list_head compound;
-	uint8_t bonds[ORIENTATION_NUM];
+	int bonds[ORIENTATION_NUM];
 	struct position pos;
 	enum element element;
 };
