+++ glyph.h
@@ -1,7 +1,6 @@
 #ifndef _GLYPH_H
 #define _GLYPH_H
 
-#include <stdint.h>
 #include "list.h"
 #include "space.h"
 
@@ -15,11 +14,11 @@ enum glyph_op {
 struct glyph {
 	struct position pos;
 	enum glyph_op op;
-	uint8_t type;
+	int type;
 	struct atom *a;
 };
 
-struct glyph *new_glyph(enum glyph_op op, uint8_t type, struct position pos);
+struct glyph *new_glyph(enum glyph_op op, int type, struct position pos);
 void delete_glyph(struct glyph *g);
 void act(struct workspace *w, struct glyph *g);
